Addr2Station/requirements.txt
requests>=2.25.1
PyYAML>=5.4.1 